requirements.txt
pyppeteer
schedule
python-dotenv